tests/wpt/web-platform-tests/signed-exchange/nested-sxg.tentative.html

<!DOCTYPE html>
<title>Nested SignedHTTPExchange</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="./resources/sxg-util.js"></script>
<body>
<script>
promise_test(async (t) => {
  const sxgUrl = get_host_info().HTTPS_ORIGIN + '/signed-exchange/resources/sxg/nested-sxg.sxg';
  const message = await openSXGInIframeAndWaitForMessage(t, sxgUrl);
  assert_equals(message.location, innerURLOrigin() + '/signed-exchange/resources/inner-url.html?fallback-from-nested-sxg');
  assert_true(message.is_fallback);
}, 'Nested SignedHTTPExchange');

</script>
</body>
